requirements.txt
# Shatably AI Requirements

# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2

# LLM Providers
google-generativeai==0.3.1
openai==1.3.5

# Image processing
Pillow==10.1.0

# Environment variables
python-dotenv==1.0.0